pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock_news_analyzer"
dynamic = ["version"]
description = "Gather, summarize, and analyze stock related news with LLM🧠"
readme = "README.md"
requires-python = ">=3.11.9"
maintainers = [
    {name = "Sungchul Kim", email = "sungchul7039@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "bs4==0.0.2",
    "langchain-community==0.3.0",
    "langchain-openai==0.2.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.3.0",
    "mypy==1.8.0",
    "tox==4.14.1",
    "pre-commit==3.7.0",
]

[tool.setuptools.dynamic]
version = {attr = "stock_news_analyzer.__version__"}

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = false  # no-untyped-def 규칙을 비활성화합니다.

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py311
isolated_build = True

[testenv]
deps = .[dev]
commands =
    ruff check .
    mypy .
    # 여기에 추가 테스트 명령어를 넣을 수 있습니다.

[testenv:pre-commit]
skip_install = true
deps = pre-commit
commands = pre-commit run --all-files
"""
